Separate WeatherList and WeatherCard
moved contents to where they belong
