+++ src/App.css
@@ -1,7 +1,7 @@
 
 /* Card */
 .MuiCard-root {
-  font-size: 30px;
+  font-size: 20px;
   height: 360px;
   width: 270px;
   margin: 10px;
