+++ src/components/Weather/WeatherList.js
@@ -1,0 +1,21 @@
+import React, { useState, useEffect } from 'react';
+import Box from '@material-ui/core/Box'
+import WeatherCard from './WeatherCard'
+import './WeatherList.css'
+
+function WeatherCards(props) {
+  function renderItems() {
+    return props.locations.map((location, index) => {
+      return (<WeatherCard location={location} key={index} index={index} isMetric={ props.isMetric }/>
+        )
+    });
+  }
+
+  return (
+    <Box className="WeatherCardsContainer" display="flex">
+      { renderItems() }
+    </Box>
+  )
+}
+
+export default WeatherCards;
